test: split webhook service test helpers
